[refactor] Clean up `ByteCodeCompileLayer.cpp`
