+++ src/jllvm/materialization/ByteCodeCompileUtils.hpp
@@ -7,8 +7,20 @@
 
 namespace jllvm
 {
-/// Returns the pointer type used for any Java reference types. This is a pointer tagged with an address space for the
-/// sake of the GC.
+/// Returns the pointer type used by the JVM for arrays of references.
+llvm::Type* arrayRefType(llvm::LLVMContext& context);
+
+/// Returns the pointer type used for any Java array types storing elements of 'elementType'.
+llvm::Type* arrayStructType(llvm::Type* elementType);
+
+/// Returns the pointer type used by the JVM for interface tables.
+llvm::Type* iTableType(llvm::LLVMContext& context);
+
+/// Returns the pointer type used by the JVM for object headers.
+llvm::Type* objectHeaderType(llvm::LLVMContext& context);
+
+/// Returns the pointer type used for all Java reference types.
+/// This is a pointer tagged with an address space for the sake of the GC.
 llvm::PointerType* referenceType(llvm::LLVMContext& context);
 
 /// Returns the corresponding LLVM type for a given Java field descriptor.
