+++ src/jllvm/materialization/CMakeLists.txt
@@ -1,8 +1,10 @@
 add_library(JLLVMMaterialization
         ByteCodeCompileLayer.cpp
+        ByteCodeCompileUtils.cpp
         ByteCodeLayer.cpp
         ByteCodeMaterializationUnit.cpp
         ByteCodeOnDemandLayer.cpp
         JNIImplementationLayer.cpp
-        ByteCodeCompileUtils.cpp ByteCodeCompileUtils.hpp)
+        CodeGenerator.cpp
+        CodeGeneratorUtils.cpp)
 target_link_libraries(JLLVMMaterialization PUBLIC JLLVMVirtualMachine JLLVMObject LLVMCore LLVMOrcJIT)
